requirements.txt
arrow==1.1.1
asgiref==3.4.1
beautifulsoup4==4.9.3
binaryornot==0.4.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
Jinja2==3.0.1
jinja2-time==0.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
Pillow==8.3.2
poyo==0.5.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
